### 2020-06-17 #2350 @whedon
[PRE REVIEW]: biorbd: A C++, Python and MATLAB library to analyze and simulate the human body biomechanics
**Submitting author:** @pariterre (<a href="http://orcid.org/0000-0002-5031-1048">Benjamin Michaud</a>)
**Repository:** <a href="https://github.com/pyomeca/biorbd" target ="_blank">https://github.com/pyomeca/biorbd</a>
**Version:** 1.3.2
