### 2020-06-17 #2350 @Kevin-Mattheus-Moerman
@pariterre we are currently looking for an editor for this submission. It may take us a bit longer than usual (see post referred to above).
In the mean time you can work on refining your paper. In particular fix those missing DOI's detected here :point_up:.  You can run `@whedon generate pdf` to update the paper and use `@whedon check references` to check the references.
