### 2020-06-17 738617f
Changing SKIP_KALMAN and STATIC_OPTIM to MODULES
