### 2020-06-17 0b8da6a
Added the capability to change segment characteristic the proper way
