### 2020-06-17 7fddd11
Added a test for this changing in segment characteristics
