+++ test/test_rigidbody.cpp
@@ -33,6 +33,7 @@ static std::string modelPathMeshEqualsMarker("models/meshsEqualMarkers.bioMod");
 static std::string modelPathForLoopConstraintTesting("models/loopConstrainedModel.bioMod");
 static std::string modelNoRoot("models/pyomecaman_freeFall.bioMod");
 static std::string modelPathForImuTesting("models/pyomecaman_withIMUs.bioMod");
+static std::string modelSimple("models/cube.bioMod");
 
 TEST(Gravity, change)
 {
@@ -903,6 +904,37 @@ TEST(Dynamics, Forward)
     for (unsigned int i = 0; i<model.nbQddot(); ++i){
         EXPECT_NEAR(static_cast<double>(QDDot(i, 0)), QDDot_expected[i], requiredPrecision);
     }
+}
+
+TEST(Dynamics, ForwardChangingMass)
+{
+    biorbd::Model model(modelSimple);
+    DECLARE_GENERALIZED_COORDINATES(Q, model);
+    DECLARE_GENERALIZED_VELOCITY(QDot, model);
+    DECLARE_GENERALIZED_TORQUE(Tau, model);
+
+    // Set to random values
+    std::vector<double> val(model.nbQ());
+    for (size_t i=0; i<val.size(); ++i){
+        val[i] = static_cast<double>(i) * 1.1;
+    }
+    FILL_VECTOR(Q, val);
+    FILL_VECTOR(QDot, val);
+    FILL_VECTOR(Tau, val);
+
+    biorbd::rigidbody::SegmentCharacteristics c(model.segment(0).characteristics());
+    c.setMass(10);
+    model.updateSegmentCharacteristics(0, c);
+
+    std::vector<double> QDDot_expected = {0.0, -9.7, 2.2};
+
+    CALL_BIORBD_FUNCTION_3ARGS(QDDot, model, ForwardDynamics, Q, QDot, Tau);
+
+    for (unsigned int i = 0; i<model.nbQddot(); ++i){
+        EXPECT_NEAR(static_cast<double>(QDDot(i, 0)), QDDot_expected[i], requiredPrecision);
+    }
+
+
 }
 
 TEST(Dynamics, ForwardDynAndExternalForces)
