+++ include/RigidBody/Segment.h
@@ -212,6 +212,14 @@ public:
     /// \param model The underlying model to update
     /// \param characteristics The new characteristics
     ///
+    /// Warning: This function may behave surpringly due to the core of RBDL. The
+    /// new characteristic values will replace everything which is attach in a fixed manner
+    /// (that is no degrees-of-freedom). So if your model has 3 segments, but only the first
+    /// one has dof (and the rest is rigidly attached to the first), then it doesn't matter
+    /// if idx is 0, 1 or 2, because RBDL considers that all these segment are 1 segment.
+    /// It is therefore expected that characteristics is the combination of mass and
+    /// inertia for these 3 segments as well.
+    ///
     void updateCharacteristics(
             biorbd::rigidbody::Joints& model,
             const biorbd::rigidbody::SegmentCharacteristics& characteristics);
