### 2020-06-25 aeea4ea
Fixed dimensions of GeneralizedTorque when using quaternions
