### 2020-06-26 dc9bda4
Added example of showing markers using vtk in Python
