### 2020-06-29 #2350 @trallard
Hello, and thanks for your patience. I will be editing your submission.

@pariterre do you have any reviewers already in mind? If not, please check our list of available reviewers <https://bit.ly/joss-reviewers> and list some names here (without tagging them with @)
