### 2020-06-29 #2350 @pariterre
Dear @trallard
Thanks for accepting being editor on my submission, this is highly appreciated!

I've had a look at the list you referred and I got 4 names that seems to fit the profile of my submission:
CELavecchia
nnadeau
abhishektha
trallard

Is this sufficient?
Thanks!
