+++ binding/python3/__init__.py
@@ -4,7 +4,7 @@ from ._version import __version__
 
 if biorbd.currentLinearAlgebraBackend() == 1:
     from casadi import Function, MX
-    def to_sx_func(name, func, *all_param):
+    def to_casadi_func(name, func, *all_param):
         mx_param = []
         for p in all_param:
             if isinstance(p, MX):
