### 2020-07-08 7e90c8a
Added shape factor buchanan to readed and tested it
