### 2020-07-10 0ba35bf
Changed the signature of stateDynamics call
