### 2020-07-16 #2350 @trallard
Dear potential reviewers,

On behalf of JOSS I would like to invite you to review the following submission:

**Title**: biorbd: A C++, Python and MATLAB library to analyze and simulate the human body biomechanics
**Repository**: https://github.com/pyomeca/biorbd
**Submitting author**: @pariterre
**Article**: <https://github.com/openjournals/joss-papers/blob/joss.02350/joss.02350/10.21105.joss.02350.pdf>

@CELavecchia are you able and willing to take this review?
@nnadeau are you able and willing to take this review?
@abhishektha are you able and willing to take this review?

Once we have at least 2 reviewers we can proceed.

FYI our reviewing guidelines can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) in case you need too have a look
