### 2020-08-03 46873d3
handling of NaN in Kalam reconstruction
