+++ src/RigidBody/KalmanReconsMarkers.cpp
@@ -11,6 +11,8 @@
 #include "RigidBody/GeneralizedAcceleration.h"
 #include "RigidBody/NodeSegment.h"
 
+#include <math.h> // isnan
+
 biorbd::rigidbody::KalmanReconsMarkers::KalmanReconsMarkers() :
     biorbd::rigidbody::KalmanRecons(),
     m_PpInitial(std::make_shared<biorbd::utils::Matrix>()),
@@ -149,7 +151,8 @@ void biorbd::rigidbody::KalmanReconsMarkers::reconstructFrame(
         // If there is a marker
         if (!Tobs(i*3).is_zero() && !Tobs(i*3+1).is_zero() && !Tobs(i*3+2).is_zero()){
 #else
-        if (Tobs(i*3)*Tobs(i*3) + Tobs(i*3+1)*Tobs(i*3+1) + Tobs(i*3+2)*Tobs(i*3+2) != 0.0){
+        if (Tobs(i*3)*Tobs(i*3) + Tobs(i*3+1)*Tobs(i*3+1) + Tobs(i*3+2)*Tobs(i*3+2) != 0.0 &&
+			!isnan(Tobs(i*3)*Tobs(i*3) + Tobs(i*3+1)*Tobs(i*3+1) + Tobs(i*3+2)*Tobs(i*3+2))){
 #endif
             H.block(i*3,0,3,*m_nbDof) = J_tp[i];
             zest.block(i*3, 0, 3, 1) = zest_tp[i];
