+++ doc/paper/paper.bib
@@ -16,11 +16,21 @@
   number = {2}
 }
 
-@article{Andersson2018,
-  title = {{{CasADi}} \textendash{} {{A}} Software Framework for Nonlinear Optimization and Optimal Control},
-  author = {Andersson, Joel A E and Gillis, Joris and Horn, Greg and Rawlings, James B and Diehl, Moritz},
-  year = {In Press, 2018},
-  journal = {Mathematical Programming Computation}
+@article{anderssonCasADiSoftwareFramework2019a,
+  title = {{{CasADi}}: A Software Framework for Nonlinear Optimization and Optimal Control},
+  shorttitle = {{{CasADi}}},
+  author = {Andersson, Joel A. E. and Gillis, Joris and Horn, Greg and Rawlings, James B. and Diehl, Moritz},
+  year = {2019},
+  month = mar,
+  volume = {11},
+  pages = {1--36},
+  issn = {1867-2957},
+  doi = {10.1007/s12532-018-0139-4},
+  abstract = {We present CasADi, an open-source software framework for numerical optimization. CasADi is a general-purpose tool that can be used to model and solve optimization problems with a large degree of flexibility, larger than what is associated with popular algebraic modeling languages such as AMPL, GAMS, JuMP or Pyomo. Of special interest are problems constrained by differential equations, i.e. optimal control problems. CasADi is written in self-contained C++, but is most conveniently used via full-featured interfaces to Python, MATLAB or Octave. Since its inception in late 2009, it has been used successfully for academic teaching as well as in applications from multiple fields, including process control, robotics and aerospace. This article gives an up-to-date and accessible introduction to the CasADi framework, which has undergone numerous design improvements over the last 7~years.},
+  file = {/home/pariterre/Zotero/storage/9AJFYRI2/Andersson et al. - 2019 - CasADi a software framework for nonlinear optimiz.pdf},
+  journal = {Mathematical Programming Computation},
+  language = {en},
+  number = {1}
 }
 
 @article{belaiseEMGmarkerTrackingOptimisation2018a,
@@ -121,7 +131,7 @@
   number = {2}
 }
 
-@article{fohannoEstimation3DKinematics2010,
+@article{fohannoEstimation3DKinematics2010b,
   title = {Estimation of the {{3D}} Kinematics in Kayak Using an Extended {{Kalman}} Filter Algorithm: A Pilot Study},
   shorttitle = {Estimation of the {{3D}} Kinematics in Kayak Using an Extended {{Kalman}} Filter Algorithm},
   author = {Fohanno, V. and Colloud, F. and Begon, M. and Lacouture, P.},
@@ -129,8 +139,9 @@
   volume = {13},
   pages = {55--56},
   publisher = {{Taylor \& Francis Journals}},
+  doi = {10.1080/10255842.2010.491958},
   abstract = {No abstract is available for this item.},
-  file = {/home/pariterre/Zotero/storage/R29NZWP6/v13y2010is1p55-56.html},
+  file = {/home/pariterre/Zotero/storage/K3LJ8WNU/v13y2010is1p55-56.html},
   journal = {Computer Methods in Biomechanics and Biomedical Engineering},
   language = {en},
   number = {S1}
@@ -232,10 +243,14 @@
 
 @book{kurfessRoboticsAutomationHandbook2018,
   title = {Robotics and Automation Handbook},
+  shorttitle = {Robotics and Automation Handbook},
   author = {Kurfess, Thomas},
   year = {2018},
   edition = {CRC press},
[…]
