+++ doc/paper/paper.bib
@@ -241,16 +241,17 @@
   number = {12}
 }
 
-@book{kurfessRoboticsAutomationHandbook2018,
-  title = {Robotics and Automation Handbook},
-  shorttitle = {Robotics and Automation Handbook},
-  author = {Kurfess, Thomas},
+@book{kurfessRoboticsAutomationHandbook2018a,
+  title = {Robotics and {{Automation Handbook}}},
+  author = {Kurfess, Thomas R.},
   year = {2018},
-  edition = {CRC press},
-  publisher = {{Thomas R. Kurfess}},
-  abstract = {As the capability and utility of robots has increased dramatically with new technology, robotic systems can perform tasks that are physically dangerous for humans, repetitive in nature, or require increased accuracy, precision, and sterile conditions to radically minimize human error. The Robotics and Automation Handbook addresses the major aspects of designing, fabricating, and enabling robotic systems and their various applications. It presents kinetic and dynamic methods for analyzing robotic systems, considering factors such as force and torque. From these analyses, the book develops several controls approaches, including servo actuation, hybrid control, and trajectory planning. Design aspects include determining specifications for a robot, determining its configuration, and utilizing sensors and actuators. The featured applications focus on how the specific difficulties are overcome in the development of the robotic system. With the ability to increase human safety and precision in applications ranging from handling hazardous materials and exploring extreme environments to manufacturing and medicine, the uses for robots are growing steadily. The Robotics and Automation Handbook provides a solid foundation for engineers and scientists interested in designing, fabricating, or utilizing robotic systems.},
-  isbn = {978-0-8493-1804-7},
-  language = {Anglais}
+  month = oct,
+  publisher = {{CRC Press}},
+  doi = {10.1201/9781315220352},
+  abstract = {As the capability and utility of robots has increased dramatically with new technology, robotic systems can perform tasks that are physically dangerous for},
+  file = {/home/pariterre/Zotero/storage/8THQNY89/Kurfess - 2018 - Robotics and Automation Handbook.pdf;/home/pariterre/Zotero/storage/XHA5WMDJ/9781315220352.html},
+  isbn = {978-1-315-22035-2},
+  language = {en}
 }
 
 @inproceedings{macchiettoMomentumControlBalance2009,
