### 2020-09-15 fad1617
Removed an unnecessary allocation
