+++ test/test_muscles.cpp
@@ -1536,10 +1536,6 @@ TEST(WrappingCylinder, unitTest)
             biorbd::utils::Vector3d(1., 1., 1.), biorbd::utils::Vector3d(1., 1., 1.), "xyz");
 
         biorbd::muscles::WrappingCylinder wrappingCylinder;
-        
-        wrappingCylinder.setLength(2.0);
-        SCALAR_TO_DOUBLE(length, wrappingCylinder.length());
-        EXPECT_NEAR(length, 2.0, requiredPrecision);
 
         wrappingCylinder.setDiameter(1.5);
         SCALAR_TO_DOUBLE(diameter, wrappingCylinder.diameter());
@@ -1554,15 +1550,15 @@ TEST(WrappingCylinder, unitTest)
         biorbd::utils::RotoTrans rt(
             biorbd::utils::Vector3d(1., 1., 1.), biorbd::utils::Vector3d(1., 1., 1.), "xyz");
 
-        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1.5, true);
-        SCALAR_TO_DOUBLE(length, wrappingCylinder.length());
-        EXPECT_NEAR(length, 1.5, requiredPrecision);
+        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true);
+        SCALAR_TO_DOUBLE(diameter, wrappingCylinder.diameter());
+        EXPECT_NEAR(diameter, 0.5, requiredPrecision);
     }
     {
         biorbd::utils::RotoTrans rt(
             biorbd::utils::Vector3d(1., 1., 1.), biorbd::utils::Vector3d(1., 1., 1.), "xyz");
 
-        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1.5, true, "name", "parentName");
+        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true, "name", "parentName");
         EXPECT_STREQ(wrappingCylinder.parent().c_str(), "parentName");
     }
     {
@@ -1600,7 +1596,7 @@ TEST(WrappingCylinder, deepCopy)
     biorbd::utils::RotoTrans rt(
         biorbd::utils::Vector3d(1, 1, 1), biorbd::utils::Vector3d(1, 1, 1), "xyz");
 
-    biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1.5, true);
+    biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true);
 
     biorbd::muscles::WrappingCylinder shallowCopy(wrappingCylinder);
     biorbd::muscles::WrappingCylinder deepCopyNow(wrappingCylinder.DeepCopy());
@@ -1608,29 +1604,29 @@ TEST(WrappingCylinder, deepCopy)
     deepCopyLater.DeepCopy(wrappingCylinder);
 
     {
-        SCALAR_TO_DOUBLE(length, wrappingCylinder.length());
-        SCALAR_TO_DOUBLE(shallowCopyLength, shallowCopy.length());
-        SCALAR_TO_DOUBLE(deepCopyNowLength, deepCopyNow.length());
-        SCALAR_TO_DOUBLE(deepCopyLaterLength, deepCopyLater.length());
-        EXPECT_NEAR(length, 1.5, requiredPrecision);
-        EXPECT_NEAR(shallowCopyLength, 1.5, requiredPrecision);
-        EXPECT_NEAR(deepCopyNowLength, 1.5, requiredPrecision);
-        EXPECT_NEAR(deepCopyLaterLength, 1.5, requiredPrecision);
[…]
