### 2020-09-28 33c6c48
Made possible to call function with native biorbd types
