+++ binding/python3/biorbd_python.i.in
@@ -384,20 +384,16 @@
 %typemap(typecheck, precedence=2100)
 biorbd::rigidbody::GeneralizedCoordinates & {
     void *argp1 = 0;
+    if(SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_biorbd__rigidbody__GeneralizedCoordinates,  0  | 0)) && argp1){
+            $1 = true;
+        }
 #ifdef BIORBD_USE_CASADI_MATH
-    if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_casadi__MX,  0  | 0)) && argp1) {
-#else
-    if (SWIG_IsOK(
-                SWIG_ConvertPtr(
-                    $input, &argp1,
-                    SWIGTYPE_p_biorbd__rigidbody__GeneralizedCoordinates,
-                    0  | 0))
-            && argp1) {
-#endif
+    else if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_casadi__MX,  0  | 0)) && argp1) {
         // Test if it is a pointer to
         // SWIGTYPE_p_biorbd__rigidbody__GeneralizedCoordinates already exists
         $1 = true;
     }
+#endif
     else if( PyArray_Check($input) ) {
         // test if it is a numpy array
         $1 = true;
@@ -408,19 +404,14 @@ biorbd::rigidbody::GeneralizedCoordinates & {
 }
 %typemap(in) biorbd::rigidbody::GeneralizedCoordinates & {
     void * argp1 = 0;
-#ifdef BIORBD_USE_CASADI_MATH
-    if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_casadi__MX,  0  | 0)) && argp1) {
-        $1 = new biorbd::rigidbody::GeneralizedCoordinates(*reinterpret_cast<casadi::MX*>(argp1));
-#else
-    if (SWIG_IsOK(
-                SWIG_ConvertPtr(
-                    $input, &argp1,
-                    SWIGTYPE_p_biorbd__rigidbody__GeneralizedCoordinates,
-                    0  | 0))
-            && argp1) {
+    if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1,SWIGTYPE_p_biorbd__rigidbody__GeneralizedCoordinates, 0  | 0)) && argp1) {
         $1 = reinterpret_cast< biorbd::rigidbody::GeneralizedCoordinates * >(argp1);
-#endif
     }
+#ifdef BIORBD_USE_CASADI_MATH
+    else if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_casadi__MX,  0  | 0)) && argp1) {
+        $1 = new biorbd::rigidbody::GeneralizedCoordinates(*reinterpret_cast<casadi::MX*>(argp1));
+    }
+#endif
     else if( PyArray_Check($input) ) {
         // Get dimensions of the data
         int        ndim     = PyArray_NDIM    ((PyArrayObject*)$input);
@@ -446,20 +437,14 @@ biorbd::rigidbody::GeneralizedCoordinates & {
 };
 %typemap(typecheck, precedence=2101) biorbd::rigidbody::GeneralizedCoordinates * {
     void *argp1 = 0;
-#ifdef BIORBD_USE_CASADI_MATH
-    if (SWIG_IsOK(SWIG_ConvertPtr($input, &argp1, SWIGTYPE_p_casadi__MX,  0  | 0)) && argp1) {
-#else
-    if (SWIG_IsOK(
-                SWIG_ConvertPtr(
[…]
