### 2020-09-28 9ca4cd9
Segmented casadi test in travis
