### 2020-09-28 697c857
Changed required CMake version for Python
