### 2020-09-29 9526e83
removed temporarily python test with casadi
