### 2020-09-29 003fcd2
Added Python binder to casadi travis
