### 2020-09-29 f74857c
upgraded cmake so it finds numpy
