### 2020-09-29 1e1b16f
Corrected path in travis after adding python3
