### 2020-09-29 ab949a0
Allowed for SpatialVector to be filled with np array and adjusted tests
