### 2020-10-03 pyomeca/biorbd#172 @abhishektha
[REVIEW] Not able to run staticOptimization example for Python
With the latest version from conda-forge I get the error:
```
Traceback (most recent call last):
  File "staticOptimization.py", line 38, in <module>
    optim = biorbd.StaticOptimization(model, Q, Qdot, Tau)
AttributeError: module 'biorbd' has no attribute 'StaticOptimization'
```

when running the staticOptimization.py Python example. I'm assuming the conda-forge version is using Eigen, which means this should be possible?

for JOSS Review https://github.com/openjournals/joss-reviews/issues/2562
