### 2020-10-04 pyomeca/biorbd#172 @pariterre
Hi @abhishektha
Are you on Windows? If so, because of conflicting version of VS of Ipopt on conda-forge, I am unable to compile with the StaticOptimization module for that particular OS (it causes a linking error) for anaconda using conda-forge.

Therefore, in order to use the StaticOptimization module on Windows, one must compile it by themselves.

If you need, I could compile a version and add it to a self hosted channel on anaconda!
