### 2020-10-04 pyomeca/biorbd#172 @abhishektha
Yes this was on Windows! I don't think you need a new version on Anaconda for the paper submission, but maybe just put a note in the documentation about it (unless I missed it)!

I'll be trying the self-compiled versions later today!
