### 2020-10-05 pyomeca/biorbd#174 @abhishektha
[REVIEW] C++ Compiling Help 
I'm assuming the CMakeLists.txt in the examples folder is supposed to help compile the example c++ scripts.
When I try to run it, I get an error:
```
CMake Error at CMakeLists.txt:21 (add_dependencies):
  add_dependencies called with incorrect number of arguments
```
I think this is due to the $BIORBD_NAME variable not being defined. What is this variable supposed to represent, and what should I set it to? Should I also be defining the locations for the biorbd_LIBRARY and biorbd_INCLUDE_DIR in the CMakeLists.txt? And should I also be defining the locations for Eigen and rbdl includes?

I alternatively tried compiling with g++, and put in the biorbd, Eigen, and rbdl include libraries in the command line. Unfortunately I got the error  "undefined reference to `biorbd::rigidbody::GeneralizedVelocity::~GeneralizedVelocity()'", repeated for multiple class:functions. Not sure if this an error on my end, or with how the class:functions are defined in biorbd.

Sorry if some of these questions are basic, I'm a little rusty on compiling my c++ scripts!

For JOSS review https://github.com/openjournals/joss-reviews/issues/2562
