### 2020-10-05 pyomeca/biorbd#174 @pariterre
Hello @abhishektha !
Sorry for the confusion! Examples are indeed examples on how to use `biorbd` but not how to compile them per se. So you have nothing special to do. When you compile `biorbd` (the main library), there is an option BUILD_EXAMPLE that is automatically set to ON. If one wants to not build the examples, they can set it to OFF. Otherwise, it will automatically build them.

So bottom line, if you already build `biorbd`, you probably have the examples already built in your build folder (`$BUILD_FOLDER/examples/$BUILD_TYPE/  where $BUILD_TYPE is the selected optimisation (usually Release, Debug or RelWithDebInfo) )

The reason there is no actual example of how to use `biorbd` outside my framework is because it is the usual workflow of CMake. That is using the command `find_package(biorbd REQUIRED)` and then linking and adding the headers via `biorbd_LIBRARIES` and `biorbd_INCLUDE_DIR`)

I hope this helps!
