+++ share/biorbdConfig.cmake.in
@@ -2,10 +2,61 @@ set(@CMAKE_PROJECT_NAME@_VERSION @PROJECT_VERSION_MAJOR@.@PROJECT_VERSION_MINOR@
 
 @PACKAGE_INIT@
 
-set_and_check(biorbd_INCLUDE_DIR "@PACKAGE_CMAKE_INSTALL_PREFIX@/@biorbd_INCLUDE_FOLDER@")
-set_and_check(biorbd_LIBRARY_DIR "@PACKAGE_CMAKE_INSTALL_PREFIX@/@biorbd_LIB_FOLDER@")
+if (biorbd_INCLUDE_DIR)
+  # Already in cache, be silent
+  set (biorbd_FIND_QUIETLY TRUE)
+endif (biorbd_INCLUDE_DIR)
 
-set(@CMAKE_PROJECT_NAME@_LIBRARY @CMAKE_PROJECT_NAME@)
+find_path (biorbd_INCLUDE_DIR "BiorbdModel.h" PATHS ${PACKAGE_PREFIX_DIR}/include/biorbd)
+find_library (biorbd_LIBRARY NAMES biorbd biorbd_debug PATHS ${PACKAGE_PREFIX_DIR}/lib/biorbd)
 
-include("${CMAKE_CURRENT_LIST_DIR}/@CMAKE_PROJECT_NAME@Targets.cmake")
+get_filename_component(biorbd_LIB_PATH ${biorbd_LIBRARY} DIRECTORY)
+get_filename_component(biorbd_LIB_NAME ${biorbd_LIBRARY} NAME_WE)
+get_filename_component(biorbd_LIB_EXTENSION ${biorbd_LIBRARY} EXT)
+
+string(REGEX MATCH "_debug" debug_flag ${biorbd_LIB_NAME})
+if (debug_flag)
+    string(REGEX REPLACE ${debug_flag} "" biorbd_LIB_NAME ${biorbd_LIB_NAME})
+endif()
+
+set(biorbd_LIBRARIES
+    ${biorbd_LIB_PATH}/${biorbd_LIB_NAME}${debug_flag}${biorbd_LIB_EXTENSION}
+    ${biorbd_LIB_PATH}/${biorbd_LIB_NAME}_utils${debug_flag}${biorbd_LIB_EXTENSION}
+    ${biorbd_LIB_PATH}/${biorbd_LIB_NAME}_rigidbody${debug_flag}${biorbd_LIB_EXTENSION}
+    ${biorbd_LIB_PATH}/${biorbd_LIB_NAME}_muscles${debug_flag}${biorbd_LIB_EXTENSION}
+    ${biorbd_LIB_PATH}/${biorbd_LIB_NAME}_actuators${debug_flag}${biorbd_LIB_EXTENSION}
+)
+
+# handle the QUIETLY and REQUIRED arguments and set biorbd_FOUND to TRUE if 
+# all listed variables are TRUE
+include (FindPackageHandleStandardArgs)
+find_package_handle_standard_args (biorbd DEFAULT_MSG 
+  biorbd_LIBRARIES
+  biorbd_INCLUDE_DIR
+)
+
+
+get_filename_component(biorbd_LIB_PATH ${biorbd_LIBRARY} DIRECTORY)
+get_filename_component(biorbd_LIB_NAME ${biorbd_LIBRARY} NAME_WE)
+get_filename_component(biorbd_LIB_EXTENSION ${biorbd_LIBRARY} EXT)
+
+string(REGEX MATCH "_debug" debug_flag ${biorbd_LIB_NAME})
+if (debug_flag)
+    string(REGEX REPLACE ${debug_flag} "" biorbd_LIB_NAME ${biorbd_LIB_NAME})
[…]
