### 2020-10-05 567331e
Added an example of usage from installed biorbd
