+++ examples/cpp_installed/README.txt
@@ -1,0 +1,1 @@
+This is an example on how to link biorbd using CMake for an installed version of biorbd
