### 2020-10-05 1710c46
Working version of the example from install
