+++ examples/cpp_installed/README.txt
@@ -1,1 +1,1 @@
-This is an example on how to link biorbd using CMake for an installed version of biorbd
+This is an example on how to link biorbd using CMake for a previously installed version of biorbd
