### 2020-10-05 cb6ed86
Making the installed example work for windows
