### 2020-10-05 e13210a
added an example in travis for the installed version of biorbd
