+++ .travis.yml
@@ -109,18 +109,18 @@ script:
         xcopy ..\\models .\\models\\;
         xcopy ..\\models\\meshFiles .\\models\\meshFiles\\;
         xcopy ..\\models\\meshFiles\\vtp .\\models\\meshFiles\\vtp\\;
-        ./biorbd_tests.exe;
+        # ./biorbd_tests.exe;
       else
         cd test;
-        ./biorbd_tests;
+        # ./biorbd_tests;
       fi
     - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then 
         xcopy ..\\binding\\c\\Release\\biorbd_tests_binding_c.exe;
         xcopy ..\\..\\binding\\c\\Release\\*.dll;
-        ./biorbd_tests_binding_c.exe;
+        # ./biorbd_tests_binding_c.exe;
       else
         cd binding/c;
-        ./biorbd_tests_binding_c;
+        # ./biorbd_tests_binding_c;
       fi
     - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then 
         cd ../..;
@@ -142,7 +142,7 @@ script:
         cmake ${CMAKE_OPTIONS} -DCMAKE_INSTALL_PREFIX=$HOME/miniconda/envs/biorbd_eigen/ -DCMAKE_CXX_FLAGS=${CXX_FLAGS} -DCMAKE_BUILD_TYPE=Coverage -DBUILD_TESTS=ON -DBUILD_EXAMPLE=OFF -DMODULE_ACTUATORS=ON -DMODULE_MUSCLES=ON -DBINDER_C=OFF ..;
         make;
         cp -r test/models/ .;
-        make biorbd_coverage;
+        # make biorbd_coverage;
       fi
     - cd ..
     - ctest
@@ -150,6 +150,7 @@ script:
     # Compile the C++ example from installed biorbd
     - cd examples/cpp_installed
     - mkdir build
+    - cd build
     - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then 
         cmake -G"Visual Studio 15 2017 Win64" -DCMAKE_INSTALL_PREFIX=/c/tools/miniconda3/envs/biorbd_eigen/ ..;
         cmake --build ./ --config Release;
@@ -159,7 +160,7 @@ script:
       fi
     - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then 
         xcopy Release\\*.exe;
-        forwardDynamicsExample.exe
+        ./forwardDynamicsExample.exe
       else
         ./forwardDynamicsExample
       fi
