### 2020-10-05 1c53c07
Added C++11 to Cmakelist and fixed typo
