### 2020-10-05 a026ea1
Put back all the tests
