### 2020-10-05 pyomeca/biorbd#174 @pariterre
I added an actual example of how to link a code to an current install of biorbd. It made me found a bug in the cmake share.
Thanks for that!
See #175
