### 2020-10-05 6605a7f
Added warning for Ipopt on Windows
