### 2020-10-05 pyomeca/biorbd#172 @pariterre
@abhishektha
Yep, good idea! It is now done thanks to #176
