+++ src/Utils/Rotation.cpp
@@ -317,8 +317,10 @@ void biorbd::utils::Rotation::checkUnitary()
 {
 #ifndef BIORBD_USE_CASADI_MATH
 #ifndef SKIP_ASSERT
-    biorbd::utils::Error::check(fabs(this->squaredNorm() - 3.) < 1e-6,
-                                "The Rotation matrix norm is not equal to one");
+    double sqrtNorm = static_cast<double>(this->squaredNorm());
+    biorbd::utils::Error::check(fabs(sqrtNorm - 3.) < 1e-6,
+                                biorbd::utils::String("The Rotation matrix square norm is equal to ")
+                                + sqrtNorm + ", but should be equal to 3");
 #endif
 #endif
 }
