### 2020-10-08 cf48570
added pair to SWIG
