### 2020-10-21 8df1792
Updated paper.bib to fix some references
References fixed:
Eigen: added all authors instead of just "and others"
Kurfess: added DOI
biorbdoptim & biorbd-viz: updated to add URLs by using the bibtex from those repositories
