+++ doc/paper/paper.bib
@@ -92,10 +92,10 @@
   series = {Lecture {{Notes}} in {{Control}} and {{Information Sciences}}}
 }
 
-@article{eigenweb,
-  title = {Eigen V3},
-  author = {Guennebaud, Ga{\"e}l and Jacob, Beno{\^i}t and others},
-  year = {2010}
+@misc{guennebaud2010eigen,
+  title={Eigen v3},
+  author={Guennebaud, Ga{\"e}l and Jacob, Beno{\^\i}t and Avery, Philip and Bachrach, Abraham and Barthelemy, Sebastien and others},
+  year={2010}
 }
 
 @inproceedings{featherstoneRobotDynamicsEquations2000,
@@ -251,6 +251,7 @@
   abstract = {As the capability and utility of robots has increased dramatically with new technology, robotic systems can perform tasks that are physically dangerous for},
   file = {/home/pariterre/Zotero/storage/8THQNY89/Kurfess - 2018 - Robotics and Automation Handbook.pdf;/home/pariterre/Zotero/storage/XHA5WMDJ/9781315220352.html},
   isbn = {978-1-315-22035-2},
+  doi = {10.1201/9781315220352},
   language = {en}
 }
 
@@ -303,18 +304,20 @@
   number = {5}
 }
 
-@article{Michaud2018biorbdoptim,
-  title = {Biorbd-Optim: {{An}} Optimal Control Framework for Biomechanical Analyses Using Biorbd},
-  author = {Michaud, Benjamin and Begon, Mickael},
-  year = {2020},
-  howpublished = {Web page}
+@misc{Michaud2018bioptim, 
+  author = {Michaud, Benjamin and Begon, Mickael}, 
+  title = {bioptim: An optimal control framework for biomechanical analyses using biorbd}, 
+  howpublished={Web page}, 
+  url = {https://github.com/pyomeca/bioptim}, 
+  year = {2020}
 }
 
-@article{Michaud2018biorbdViz,
-  title = {Biorbd-Viz: {{A}} Vizualization Python Toolbox for Biorbd},
-  author = {Michaud, Benjamin and Begon, Mickael},
-  year = {2018},
-  howpublished = {Web page}
+@misc{Michaud2018biorbdViz,
+    author = {Michaud, Benjamin and Begon, Mickael},
+    title = {bioviz: A vizualization python toolbox for biorbd},
+    howpublished={Web page},
+    url = {https://github.com/pyomeca/bioviz},
+    year = {2018}
 }
 
 @article{moissenetOptimizationMethodTracking2019a,
