### 2020-10-21 pyomeca/biorbd#180 @abhishektha
[REVIEW] Fix reference on JOSS Paper
I can't seem to find this article that you have in your references, do you have more information as to where it was presented?

Desmyttere, G., Hajizadeh, M., Bleau, J., & Begon, M. (2019). Foot orthosis with add-onrearfoot posting can alter foot kinematics. In.

For JOSS Review https://github.com/openjournals/joss-reviews/issues/2562#
