### 2020-10-21 pyomeca/biorbd#180 @pariterre
Hi there!
That is an InProceedings from the ASB conference (https://www.researchgate.net/publication/335025832_Foot_orthosis_with_add-on_rearfoot_posting_can_alter_foot_kinematics)
However, since, the paper has actually been publish, so I changed the reference for the actual paper! Thanks for noticing!
