+++ doc/paper/paper.bib
@@ -67,6 +67,22 @@
   series = {{{SIMS}} 2004}
 }
 
+@article{desmyttereEffect3DPrinted2020,
+  title = {Effect of {{3D}} Printed Foot Orthoses Stiffness and Design on Foot Kinematics and Plantar Pressures in Healthy People},
+  author = {Desmyttere, Gauthier and Leteneur, S{\'e}bastien and Hajizadeh, Maryam and Bleau, Jacinte and Begon, Mickael},
+  year = {2020},
+  month = sep,
+  volume = {81},
+  pages = {247--253},
+  issn = {0966-6362},
+  doi = {10.1016/j.gaitpost.2020.07.146},
+  abstract = {Background Foot orthoses (FOs) have been widely prescribed to alter various lower limb disorders. FOs' geometrical design and material properties have been shown to influence their impact on foot biomechanics. New technologies such as 3D printing provide the potential to produce custom shapes and add functionalities to FOs by adding extra-components. Research question The purpose of this study was to determine the effect of 3D printed FOs stiffness and newly design postings on foot kinematics and plantar pressures in healthy people. Methods Two pairs of {$\frac{3}{4}$} length prefabricated 3D printed FOs were administered to 15 healthy participants with normal foot posture. FOs were of different stiffness and were designed so that extra-components, innovative flat postings, could be inserted at the rearfoot. In-shoe multi-segment foot kinematics as well as plantar pressures were recorded while participants walked on a treadmill. One-way ANOVAs using statistical non-parametric mapping were performed to estimate the effect of FOs stiffness and then the addition of postings during the stance phase of walking. Results Increasing FOs stiffness altered frontal and transverse plane foot kinematics, especially by further reducing rearfoot eversion and increasing the rearfoot abduction. Postings had notable effect on rearfoot frontal plane kinematics, by enhancing FOs effect. Looking at plantar pressures, wearing FOs was associated with a shift of the loads from the rearfoot to the midfoot region. Higher peak pressures under the rearfoot and midfoot (up to +31.7 \%) were also observed when increasing the stiffness of the FOs. Significance 3D printing techniques offer a wide range of possibilities in terms of material properties and design, providing clinicians the opportunity to administer FOs that could be modulated according to pathologies as well as during the treatment by adding extra-components. Further studies including people presenting musculoskeletal disorders are required.},
+  file = {/home/pariterre/Zotero/storage/LD7HY8LH/Desmyttere et al. - 2020 - Effect of 3D printed foot orthoses stiffness and d.pdf;/home/pariterre/Zotero/storage/2BW4XDZN/S0966636220304926.html},
+  journal = {Gait \& Posture},
+  keywords = {3D printing,Foot kinematics,Foot orthoses,Gait,Plantar pressures},
+  language = {en}
+}
+
 @inproceedings{desmyttereFootOrthosisAddon2019,
   title = {Foot Orthosis with Add-on Rearfoot Posting Can Alter Foot Kinematics},
   author = {Desmyttere, Gauthier and Hajizadeh, Maryam and Bleau, Jacinte and Begon, Mickael},
