### 2020-10-21 pyomeca/biorbd#180 @pariterre
(See #179 )
