+++ doc/paper/paper.bib
@@ -328,7 +328,7 @@
   year = {2020}
 }
 
-@misc{Michaud2018biorbdViz,
+@misc{Michaud2018biorviz,
     author = {Michaud, Benjamin and Begon, Mickael},
     title = {bioviz: A vizualization python toolbox for biorbd},
     howpublished={Web page},
