### 2020-10-22 #2562 @abhishektha
@trallard I can recommend biorbd for publication in JOSS. It is a very useful alternative to established programs like OpenSim for biomechanics analysis. biorbd easily calculates both forward and inverse kinematics and dynamics from given inputs using well-optimized C++ code. MATLAB and Python libraries allow for easy analysis in biorbd.
