### 2020-10-22 f245924
Add scrit to plot surface representing the relation of q qdot tau
