### 2020-10-22 ee9dc80
Updated test_actuator for non negative r for gauss3p
