+++ test/test_actuators.cpp
@@ -185,8 +185,8 @@ TEST(Actuators, jointTorqueFromAllTypesOfActuators){
 
             FILL_VECTOR(actuatorActivations, act_val);
 
-            std::vector<double> tauExpected = {-2.5, -216.9126638415241, -28.111094840325926, -160.06339366097643};
-            std::vector<double> torqueMaxExpected = {5, 302.7521746539316, 57.907864248091514, 320.12678732195286};
+            std::vector<double> tauExpected = {-2.5, -71.180973314287982, -28.111094840325926, -160.06339366097643};
+            std::vector<double> torqueMaxExpected = {5, 99.349637237541288, 57.907864248091514, 320.12678732195286};
 
             CALL_BIORBD_FUNCTION_3ARGS(tau, model, torque, actuatorActivations, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
@@ -252,8 +252,8 @@ TEST(Actuators, jointTorqueFromAllTypesOfActuators){
             }
             FILL_VECTOR(actuatorActivations, act_val);
 
-            std::vector<double> tauExpected = {-2.5, -151.3760873269658, -28.953932124045757, -160.06339366097643};
-            std::vector<double> torqueMaxExpected = {5, 433.8253276830482, 56.22218968065185, 320.12678732195286};
+            std::vector<double> tauExpected = {-2.5, -49.674818618770644, -28.953932124045757, -160.06339366097643};
+            std::vector<double> torqueMaxExpected = {5, 142.36194662857596, 56.22218968065185, 320.12678732195286};
 
             CALL_BIORBD_FUNCTION_3ARGS(tau, model, torque, actuatorActivations, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
