### 2020-10-22 6f9592a
Made surface_max_torque_actuator work for eigen
