### 2020-10-26 #2562 @trallard
Thanks for your review and contribution to JOSS @abhishektha this is much appreciated

---
@pariterre your paper is needing an explicit [Statemen t of Need](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) section.
I know this is pretty much covered in the Summary section, but we are now requesting for this to be a separate section for al JOSS papers
