### 2020-10-26 #2562 @pariterre
@trallard
Hi! Thanks for reviewing, that is really appreciated. There is a section called "The need for `biorbd`" Do you mean that I must add another one or just that the section's name should be "Statement of need"?
