### 2020-10-27 3c52a9f
Get one marker when providing a marker name
