### 2020-10-27 596638c
Added a setPosition for Vector3d and solved a SegFault when using it in casadi
