### 2020-10-30 a5797a7
Reverted change in getting imu names
