+++ src/RigidBody/Markers.cpp
@@ -5,7 +5,6 @@
 #include <rbdl/Kinematics.h>
 #include "Utils/String.h"
 #include "Utils/Matrix.h"
-#include "Utils/Error.h"
 #include "RigidBody/GeneralizedCoordinates.h"
 #include "RigidBody/GeneralizedVelocity.h"
 #include "RigidBody/Joints.h"
@@ -64,13 +63,15 @@ const biorbd::rigidbody::NodeSegment &biorbd::rigidbody::Markers::marker(
     return (*m_marks)[idx];
 }
 
-const biorbd::rigidbody::NodeSegment& biorbd::rigidbody::Markers::marker(
+std::vector<biorbd::rigidbody::NodeSegment> biorbd::rigidbody::Markers::marker(
         const biorbd::utils::String& name) const
 {
+    std::vector<biorbd::rigidbody::NodeSegment> pos;
     for (unsigned int i=0; i<nbMarkers(); ++i) // Go through all the markers and select the right ones
         if (!marker(i).parent().compare(name))
-            return marker(i);
-    biorbd::utils::Error::raise("Marker " + name + " not found");
+            pos.push_back(marker(i));
+
+    return pos;
 }
 
 // Return a marker
