### 2020-10-30 3b2b25a
Moved pyomecaman with IMU in a specific folder
