### 2020-10-30 d8462b1
Fixed a bug when readin customRT and added negative tests
