### 2020-10-31 fd53f28
Force not compiling when using CasADi and c or matlab binders
