+++ binding/CMakeLists.txt
@@ -11,12 +11,10 @@ option(BINDER_MATLAB "Build Matlab module" OFF)
 
 if (BIORBD_USE_CASADI_MATH)
     if(BINDER_C)
-        message("Casadi option is not compatible with BINDER_C")
-        set(BINDER_C OFF)
+        message(FATAL_ERROR "Casadi option is not compatible with BINDER_C")
     endif(BINDER_C)
     if(BINDER_MATLAB)
-        message("Casadi option is not compatible with BINDER_MATLAB")
-        set(BINDER_MATLAB OFF)
+        message(FATAL_ERROR "Casadi option is not compatible with BINDER_MATLAB")
     endif(BINDER_MATLAB)
 endif(BIORBD_USE_CASADI_MATH)
 
