### 2020-10-31 eab108a
Fixed the tests for c binder and windows (copy files)
