### 2020-11-06 dd221c3
Add SigmoidGauss3p to Actuators
