### 2020-11-06 ec10017
Adding SigmoidGauss3p where it was missing for compilation
