+++ src/Actuators/Actuators.cpp
@@ -10,6 +10,7 @@
 #include "Actuators/Actuator.h"
 #include "Actuators/ActuatorGauss3p.h"
 #include "Actuators/ActuatorGauss6p.h"
+#include "Actuators/ActuatorSigmoidGauss3p.h"
 #include "Actuators/ActuatorConstant.h"
 #include "Actuators/ActuatorLinear.h"
 
@@ -52,6 +53,9 @@ void biorbd::actuator::Actuators::DeepCopy(const biorbd::actuator::Actuators &ot
         else if ((*other.m_all)[i].first->type() == biorbd::actuator::TYPE::GAUSS6P)
             (*m_all)[i].first = std::make_shared<biorbd::actuator::ActuatorGauss6p>(
                     static_cast<const biorbd::actuator::ActuatorGauss6p&>( *(*other.m_all)[i].first) );
+        else if ((*other.m_all)[i].first->type() == biorbd::actuator::TYPE::SIGMOIDGAUSS3P)
+            (*m_all)[i].first = std::make_shared<biorbd::actuator::ActuatorSigmoidGauss3p>(
+                    static_cast<const biorbd::actuator::ActuatorSigmoidGauss3p&>( *(*other.m_all)[i].first) );
         else
             biorbd::utils::Error::raise("Actuator " + biorbd::utils::String(
                                             biorbd::actuator::TYPE_toStr((*other.m_all)[i].first->type()))
@@ -68,6 +72,9 @@ void biorbd::actuator::Actuators::DeepCopy(const biorbd::actuator::Actuators &ot
         else if ((*other.m_all)[i].second->type() == biorbd::actuator::TYPE::GAUSS6P)
             (*m_all)[i].second = std::make_shared<biorbd::actuator::ActuatorGauss6p>(
                     static_cast<const biorbd::actuator::ActuatorGauss6p&>( *(*other.m_all)[i].second) );
+        else if ((*other.m_all)[i].second->type() == biorbd::actuator::TYPE::SIGMOIDGAUSS3P)
+            (*m_all)[i].second = std::make_shared<biorbd::actuator::ActuatorSigmoidGauss3p>(
+                    static_cast<const biorbd::actuator::ActuatorSigmoidGauss3p&>( *(*other.m_all)[i].second) );
         else
             biorbd::utils::Error::raise("Actuator " + biorbd::utils::String(
                                             biorbd::actuator::TYPE_toStr((*other.m_all)[i].second->type()))
@@ -144,6 +151,17 @@ void biorbd::actuator::Actuators::addActuator(const biorbd::actuator::Actuator &
         }
         else{
             (*m_all)[idx].second = std::make_shared<biorbd::actuator::ActuatorGauss6p>(static_cast<const biorbd::actuator::ActuatorGauss6p&>(act));
+            (*m_isDofSet)[idx*2+1] = true;
+        }
+        return;
+    }
+    else if (act.type() == biorbd::actuator::TYPE::SIGMOIDGAUSS3P){
+        if (act.direction() == 1){
+            (*m_all)[idx].first = std::make_shared<biorbd::actuator::ActuatorSigmoidGauss3p>(static_cast<const biorbd::actuator::ActuatorSigmoidGauss3p&>(act));
+            (*m_isDofSet)[idx*2] = true;
+        }
+        else {
+            (*m_all)[idx].second = std::make_shared<biorbd::actuator::ActuatorSigmoidGauss3p>(static_cast<const biorbd::actuator::ActuatorSigmoidGauss3p&>(act));
             (*m_isDofSet)[idx*2+1] = true;
         }
         return;
@@ -251,6 +269,8 @@ biorbd::actuator::Actuators::torqueMax(
[…]
