### 2020-11-06 4db0e7b
Add torqueMax of ActuatorSigmoidGauss3p
