### 2020-11-06 4f64157
Adding dereferenciation of pointers in torqueMax of ActuatorSigmoidGauss3p
