+++ src/Actuators/ActuatorSigmoidGauss3p.cpp
@@ -100,7 +100,11 @@ biorbd::utils::Scalar biorbd::actuator::ActuatorSigmoidGauss3p::torqueMax(
     biorbd::utils::Scalar pos(Q[*m_dofIdx] * 180/M_PI);
     biorbd::utils::Scalar speed(Qdot[*m_dofIdx] * 180/M_PI);
 
-    return (m_theta / (1 + exp(lamda * m_speed)) + m_offset) * exp(-(m_qopt * pos) * (m_qopt * pos) / (2 * m_r * m_r));
+    // Getting Tmax of Guass3p from Sigmoid
+    biorbd::utils::scalar Tmax(*m_theta / (1 + exp(*m_lamda * speed)) + *m_offset);
+
+    // Calculation of the max torque
+    return Tmax * exp(-(*m_qopt * pos) * (*m_qopt * pos) / (2 * *m_r * *m_r));
 }
 
 void biorbd::actuator::ActuatorSigmoidGauss3p::setType(){
