### 2020-11-06 c8bd244
Updated comment in ActuatorSigmoidGauss3p
