### 2020-11-06 b0cc6b3
Add test for ActuatorSigmoidGauss3p
