### 2020-11-06 438dec4
Corrected test sigmoidgauss3p
