+++ test/test_actuators.cpp
@@ -284,14 +284,14 @@ TEST(ActuatorSigmoidGauss3p, torqueMax){
     DECLARE_GENERALIZED_COORDINATES(Q, model);
     DECLARE_GENERALIZED_VELOCITY(QDot, model);
 
-    biorbd::actuator::ActuatorSigmoidGauss3p sigmoid_gauss3p_torque_act(312.0781, 0.01, 3.2703, 56.4021, -25.6939);
+    biorbd::actuator::ActuatorSigmoidGauss3p sigmoid_gauss3p_torque_act(1, 312.0781, 0.01, 3.2703, 56.4021, -25.6939, 0);
     std::vector<double> Q_val(model.nbQ());
     for (size_t i=0; i<Q_val.size(); ++i){
         Q_val[i] = 1.1;
     }
     FILL_VECTOR(Q, Q_val);
     {
-        std::vector<double> QDot_val(model.nbQdot);
+        std::vector<double> QDot_val(model.nbQdot());
         for (size_t i=0; i<QDot_val.size(); ++i){
             QDot_val[i] = 10;
         }
@@ -313,7 +313,7 @@ TEST(ActuatorSigmoidGauss3p, torqueMax){
 
         double torqueMaxExpected(149.2850);
         FILL_VECTOR(QDot, QDot_val);
-        CALL_BIORBD_FUNCTION_2ARGS(torqueMaxVal, gauss3p_torque_act, torqueMax, Q, QDot);
+        CALL_BIORBD_FUNCTION_2ARGS(torqueMaxVal, sigmoid_gauss3p_torque_act, torqueMax, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
         EXPECT_NEAR(static_cast<double>(torqueMaxVal(0, 0)), torqueMaxExpected, requiredPrecision);
 #else
