### 2020-11-06 e23e10b
Corrected typo in test for ActuatorSigmoidGauss3p
