### 2020-11-07 a392044
Added bunch of updateKin when using Casadi
