### 2020-11-08 cf447cc
Removed Kalman in CasADi
