+++ src/RigidBody/CMakeLists.txt
@@ -19,6 +19,9 @@ set(SRC_LIST_MODULE
     "${CMAKE_CURRENT_SOURCE_DIR}/RotoTransNodes.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/MeshFace.cpp"
 )
+if (BIORBD_USE_CASADI_MATH AND MODULE_KALMAN)
+    message(FATAL_ERROR "Casadi and Kalman cannot be used alongside, please turn off MODULE_KALMAN or use another backend for linear algebra")
+endif()
 if (MODULE_KALMAN)
     list(APPEND SRC_LIST_MODULE
         "${CMAKE_CURRENT_SOURCE_DIR}/KalmanRecons.cpp"
