+++ src/RigidBody/CMakeLists.txt
@@ -20,7 +20,8 @@ set(SRC_LIST_MODULE
     "${CMAKE_CURRENT_SOURCE_DIR}/MeshFace.cpp"
 )
 if (BIORBD_USE_CASADI_MATH AND MODULE_KALMAN)
-    message(FATAL_ERROR "Casadi and Kalman cannot be used alongside, please turn off MODULE_KALMAN or use another backend for linear algebra")
+    message(WARNING "Casadi and Kalman cannot be used alongside, MODULE_KALMAN is automatically set to OFF")
+    set(MODULE_KALMAN OFF CACHE BOOL "" FORCE)
 endif()
 if (MODULE_KALMAN)
     list(APPEND SRC_LIST_MODULE
