+++ CMakeLists.txt
@@ -71,27 +71,29 @@ else()
     )
 endif()
 
+option(MODULE_STATIC_OPTIM "If Static optimization should be compiled" ON)
 if (IPOPT_FOUND)
-    option(MODULE_STATIC_OPTIM "If Static optimization should be compiled" ON)
-    if (BIORBD_USE_CASADI_MATH)
+    if (BIORBD_USE_CASADI_MATH AND MODULE_STATIC_OPTIM)
         message(WARNING "Static optimization can't be compiled along with Casadi")
         set(MODULE_STATIC_OPTIM OFF CACHE BOOL "" FORCE)
     endif()
 else()
     set(IPOPT_INCLUDE_DIR "")
     set(IPOPT_LIBRARY "")
-    message(WARNING "Static optimization can't be compiled since Ipopt was not found")
-    set(MODULE_STATIC_OPTIM OFF CACHE BOOL "" FORCE)
+    if (MODULE_STATIC_OPTIM)
+        message(WARNING "Static optimization can't be compiled since Ipopt was not found")
+        set(MODULE_STATIC_OPTIM OFF CACHE BOOL "" FORCE)
+    endif()
 endif()
 
-if (TinyXML_FOUND)
-    option(MODULE_VTP_FILES_READER
-        "If reader for geometry vtp files from opensim should be compiled" ON)
-else()
+option(MODULE_VTP_FILES_READER "If reader for geometry vtp files from opensim should be compiled" ON)
+if (NOT TinyXML_FOUND)
     set(TinyXML_INCLUDE_DIR "")
     set(TinyXML_LIBRARY "")
-    message(WARNING "VTP cannot be used since TinyXML was not found")
-    set(MODULE_VTP_FILES_READER OFF CACHE BOOL "" FORCE)
+    if (MODULE_VTP_FILES_READER)
+        message(WARNING "VTP cannot be used since TinyXML was not found")
+        set(MODULE_VTP_FILES_READER OFF CACHE BOOL "" FORCE)
+    endif()
 endif()
 
 
