### 2020-11-09 0bd1608
Corrected TorqueMax for actuatorsigmoidgauss3p, and adapted test
