+++ src/Actuators/ActuatorSigmoidGauss3p.cpp
@@ -104,7 +104,7 @@ biorbd::utils::Scalar biorbd::actuator::ActuatorSigmoidGauss3p::torqueMax(
     biorbd::utils::Scalar Tmax(*m_theta / (1 + exp(*m_lambda * speed)) + *m_offset);
 
     // Calculation of the max torque
-    return Tmax * exp(-(*m_qopt * pos) * (*m_qopt * pos) / (2 * *m_r * *m_r));
+    return Tmax * exp(-(*m_qopt - pos) * (*m_qopt - pos) / (2 * *m_r * *m_r));
 }
 
 void biorbd::actuator::ActuatorSigmoidGauss3p::setType(){
