+++ test/test_actuators.cpp
@@ -156,23 +156,23 @@ TEST(Actuators, jointTorqueFromAllTypesOfActuators){
             }
             FILL_VECTOR(actuatorActivations, act_val);
 
-            std::vector<double> tauExpected = {5, 45.35073139870569, 24.688715372701648, 162.56339366097643};
-            std::vector<double> torqueMaxExpected = {10, 90.701462797411381, 49.377430745403295, 325.12678732195286};
+            std::vector<double> tauExpected = {5, 45.35073139870569, 24.688715372701648, 162.56339366097643, 16.208968606473185};
+            std::vector<double> torqueMaxExpected = {10, 90.701462797411381, 49.377430745403295, 325.12678732195286, 32.41793721294637};
 
             CALL_BIORBD_FUNCTION_3ARGS(tau, model, torque, actuatorActivations, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
-            EXPECT_NEAR(tau.size().first, 4, requiredPrecision);
+            EXPECT_NEAR(tau.size().first, 5, requiredPrecision);
 #else
-            EXPECT_NEAR(tau.size(), 4, requiredPrecision);
+            EXPECT_NEAR(tau.size(), 5, requiredPrecision);
 #endif
             for (size_t i=0; i<model.nbGeneralizedTorque(); ++i){
                  EXPECT_NEAR(static_cast<double>(tau(i, 0)), tauExpected[i], requiredPrecision);
             }
             CALL_BIORBD_FUNCTION_3ARGS(torqueMax, model, torqueMax, actuatorActivations, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
-            EXPECT_NEAR(torqueMax.size().first, 4, requiredPrecision);
+            EXPECT_NEAR(torqueMax.size().first, 5, requiredPrecision);
 #else
-            EXPECT_NEAR(torqueMax.size(), 4, requiredPrecision);
+            EXPECT_NEAR(torqueMax.size(), 5, requiredPrecision);
 #endif
             for (size_t i=0; i<model.nbGeneralizedTorque(); ++i){
                  EXPECT_NEAR(static_cast<double>(torqueMax(i, 0)), torqueMaxExpected[i], requiredPrecision);
@@ -186,23 +186,23 @@ TEST(Actuators, jointTorqueFromAllTypesOfActuators){
 
             FILL_VECTOR(actuatorActivations, act_val);
 
-            std::vector<double> tauExpected = {-2.5, -71.180973314287982, -28.111094840325926, -160.06339366097643};
-            std::vector<double> torqueMaxExpected = {5, 99.349637237541288, 57.907864248091514, 320.12678732195286};
+            std::vector<double> tauExpected = {-2.5, -71.180973314287982, -28.111094840325926, -160.06339366097643, -30.025156051078717};
+            std::vector<double> torqueMaxExpected = {5, 99.349637237541288, 57.907864248091514, 320.12678732195286, 32.41793721294637};
 
             CALL_BIORBD_FUNCTION_3ARGS(tau, model, torque, actuatorActivations, Q, QDot);
 #ifdef BIORBD_USE_CASADI_MATH
-            EXPECT_NEAR(tau.size().first, 4, requiredPrecision);
+            EXPECT_NEAR(tau.size().first, 5, requiredPrecision);
 #else
-            EXPECT_NEAR(tau.size(), 4, requiredPrecision);
+            EXPECT_NEAR(tau.size(), 5, requiredPrecision);
 #endif
[…]
