### 2020-11-09 12f67bb
Fixed update muscles
