### 2020-11-09 ffd0434
Added a test for new to_casadi_func, made all tests using it and blacked
