+++ test/binding/Python3/test_binder_python_rigidbody.py
@@ -1,20 +1,16 @@
 """
 Test for file IO
 """
-from pathlib import Path
-
 import numpy as np
-import pytest
-
 import biorbd
 
 
 def test_load_model():
-    biorbd.Model("../../models/pyomecaman.bioMod")
+    biorbd.Model("pyomecaman.bioMod")
 
 
 def test_dof_ranges():
-    m = biorbd.Model("../../models/pyomecaman.bioMod")
+    m = biorbd.Model("pyomecaman.bioMod")
     pi = 3.14159265358979323846
 
     # Pelvis
@@ -72,28 +68,20 @@ def test_dof_ranges():
 
 
 def test_forward_dynamics():
-    m = biorbd.Model("../../models/pyomecaman_withActuators.bioMod")
+    m = biorbd.Model("pyomecaman_withActuators.bioMod")
 
     q = np.array([i * 1.1 for i in range(m.nbQ())])
     qdot = np.array([i * 1.1 for i in range(m.nbQ())])
     tau = np.array([i * 1.1 for i in range(m.nbQ())])
 
-
     if biorbd.currentLinearAlgebraBackend() == 1:
         # If CasADi backend is used
-        from casadi import Function, MX
+        from casadi import MX
 
         q_sym = MX.sym("q", m.nbQ(), 1)
         qdot_sym = MX.sym("qdot", m.nbQdot(), 1)
         tau_sym = MX.sym("tau", m.nbGeneralizedTorque(), 1)
-
-        ForwardDynamics = Function(
-            "ForwardDynamics",
-            [q_sym, qdot_sym, tau_sym],
-            [m.ForwardDynamics(q_sym, qdot_sym, tau_sym).to_mx()],
-            ["q_sym", "qdot_sym", "tau_sym"],
-            ["qddot"],
-        ).expand()
+        ForwardDynamics = biorbd.to_casadi_func("ForwardDynamics", m.ForwardDynamics, q_sym, qdot_sym, tau_sym)
 
         qddot = ForwardDynamics(q, qdot, tau)
         qddot = np.array(qddot)[:, 0]
@@ -101,6 +89,8 @@ def test_forward_dynamics():
     elif biorbd.currentLinearAlgebraBackend() == 0:
         # if Eigen backend is used
         qddot = m.ForwardDynamics(q, qdot, tau).to_array()
+    else:
+        raise NotImplementedError("Backend not implemented in test")
 
     qddot_expected = np.array(
         [
@@ -123,22 +113,22 @@ def test_forward_dynamics():
 
 
 def test_forward_dynamics_with_external_forces():
-    m = biorbd.Model("../../models/pyomecaman_withActuators.bioMod")
+    m = biorbd.Model("pyomecaman_withActuators.bioMod")
 
     q = np.array([i * 1.1 for i in range(m.nbQ())])
     qdot = np.array([i * 1.1 for i in range(m.nbQ())])
     tau = np.array([i * 1.1 for i in range(m.nbQ())])
 
-
     # With external forces
     if biorbd.currentLinearAlgebraBackend() == 1:
-        from casadi import Function, MX
+        from casadi import MX
+
         sv1 = MX((11.1, 22.2, 33.3, 44.4, 55.5, 66.6))
         sv2 = MX((11.1 * 2, 22.2 * 2, 33.3 * 2, 44.4 * 2, 55.5 * 2, 66.6 * 2))
     else:
         sv1 = np.array((11.1, 22.2, 33.3, 44.4, 55.5, 66.6))
         sv2 = np.array((11.1 * 2, 22.2 * 2, 33.3 * 2, 44.4 * 2, 55.5 * 2, 66.6 * 2))
-    
[…]
