### 2020-12-01 b358fd0
Added unittest for imu_to_array in Python
