### 2020-12-08 e893d1c
Added interface to marker acceleration
