### 2020-12-15 #2562 @arfon
> Hi! Thanks for reviewing, that is really appreciated. There is a section called "The need for biorbd" Do you mean that I must add another one or just that the section's name should be "Statement of need"?

@pariterre - I would suggest renaming this section _Statement of need_. From our submission guidelines:

> A clear Statement of Need that illustrates the research purpose of the software.
