### 2020-12-15 56d02b4
Added the capability to convert torque graph to degrees or not
