+++ doc/paper/paper.md
@@ -31,7 +31,7 @@ To meet the needs of the biomechanics community, Python and MATLAB binders are p
 As a result, `biorbd` can elegantly be implemented to common workflows of researchers without compromising the required speed.
 
 Finally, biomechanical data are often multidimensional and almost always time-dependent which can be challenging to visualize.
-To help with that, `biorbd-viz` [@Michaud2018biorviz], a Python visualizer, was purposely designed.
+To help with that, `bioviz` [@Michaud2018bioviz], a Python visualizer, was purposely designed.
 This visualizer allows animating the model, record videos, and, for models that include muscles, plot muscular outputs against various features of the movement.
 
 # A `biorbd` overview, the inverse and direct flow
@@ -40,7 +40,7 @@ Briefly, the former uses measurements from a movement (e.g., markers) and infers
 
 ## Inverse flow
 *Inverse kinematics*: Estimates the generalized coordinates ($q$)—i.e., the body kinematics—from body sensor measurements (e.g., markers, IMU, etc.).
-The main algorithm implemented is the Extended Kalman Filter [@fohannoEstimation3DKinematics2010b] which by design facilitates the merging of multiple data sources and takes care of missing data.
+The main algorithm implemented is the Extended Kalman Filter [@fohannoEstimation3DKinematics2010a] which by design facilitates the merging of multiple data sources and takes care of missing data.
 
 *Inverse dynamics*: Estimates the generalized forces ($\tau$) producing a given generalized acceleration ($\ddot{q}$) (the second time derivative of $q$):
 $$
@@ -48,8 +48,8 @@ $$
 $$
 where $\dot{q}$ is the generalized velocities, $M(q)$ is the mass matrix and $N(q, \dot{q})$ are the nonlinear effects.
 
-*Static optimization*: Estimates the muscle activations ($\alpha$) producing a given $\tau$ [@andersonStaticDynamicOptimization2001b].
-It minimizes the muscle activation *p*-norm ($p$ usually being $2$) that matches a given $\tau$ using nonlinear optimization (Ipopt [@wachterImplementationInteriorpointFilter2006a]).
+*Static optimization*: Estimates the muscle activations ($\alpha$) producing a given $\tau$ [@andersonStaticDynamicOptimization2001].
+It minimizes the muscle activation *p*-norm ($p$ usually being $2$) that matches a given $\tau$ using nonlinear optimization (Ipopt [@wachterImplementationInteriorpointFilter2006]).
 $$
 \begin{aligned}
     & \underset{\alpha \in \mathbb{R}^m}{\text{minimize}}
@@ -81,22 +81,22 @@ All the forward dynamics implemented in `RBDL` [@felisRBDLEfficientRigidbody2017
 
 # The dependencies
 `biorbd` takes advantage of efficient back ends, especially  the `RBDL` and `CasADi` libraries.
-`RBDL`, written by Martin Feliz [@felisRBDLEfficientRigidbody2017], implements Featherstone equations of spatial geometry [@featherstoneRobotDynamicsEquations2000], successfully used in the field of robotics [@macchiettoMomentumControlBalance2009; @diehlFastDirectMultiple2006; @kurfessRoboticsAutomationHandbook2018a]. 
+`RBDL`, written by Martin Feliz [@felisRBDLEfficientRigidbody2017], implements Featherstone equations of spatial geometry [@featherstoneRobotDynamicsEquations2000], successfully used in the field of robotics [@macchiettoMomentumControlBalance2009; @diehlFastDirectMultiple2006; @kurfessRoboticsAutomationHandbook2018]. 
 `RBDL` provides the computational core for body dynamics.
 `biorbd` extends `RBDL` by giving commonly used biomechanics nomenclature, and by adding biomechanical modules, amongst others. 
[…]
