+++ doc/paper/paper.bib
@@ -1,5 +1,5 @@
 
-@article{andersonStaticDynamicOptimization2001b,
+@article{andersonStaticDynamicOptimization2001,
   title = {Static and Dynamic Optimization Solutions for Gait Are Practically Equivalent},
   author = {Anderson, Frank C. and Pandy, Marcus G.},
   year = {2001},
@@ -16,7 +16,7 @@
   number = {2}
 }
 
-@article{anderssonCasADiSoftwareFramework2019a,
+@article{anderssonCasADiSoftwareFramework2019,
   title = {{{CasADi}}: A Software Framework for Nonlinear Optimization and Optimal Control},
   shorttitle = {{{CasADi}}},
   author = {Andersson, Joel A. E. and Gillis, Joris and Horn, Greg and Rawlings, James B. and Diehl, Moritz},
@@ -33,7 +33,7 @@
   number = {1}
 }
 
-@article{belaiseEMGmarkerTrackingOptimisation2018a,
+@article{belaiseEMGmarkerTrackingOptimisation2018,
   title = {An {{EMG}}-Marker Tracking Optimisation Method for Estimating Muscle Forces},
   author = {B{\'e}laise, Colombe and Dal Maso, Fabien and Michaud, Benjamin and Mombaur, Katja and Begon, Micka{\"e}l},
   year = {2018},
@@ -108,12 +108,6 @@
   series = {Lecture {{Notes}} in {{Control}} and {{Information Sciences}}}
 }
 
-@misc{guennebaud2010eigen,
-  title={Eigen v3},
-  author={Guennebaud, Ga{\"e}l and Jacob, Beno{\^\i}t and Avery, Philip and Bachrach, Abraham and Barthelemy, Sebastien and others},
-  year={2010}
-}
-
 @inproceedings{featherstoneRobotDynamicsEquations2000,
   title = {Robot Dynamics: Equations and Algorithms},
   shorttitle = {Robot Dynamics},
@@ -147,7 +141,7 @@
   number = {2}
 }
 
-@article{fohannoEstimation3DKinematics2010b,
+@article{fohannoEstimation3DKinematics2010a,
   title = {Estimation of the {{3D}} Kinematics in Kayak Using an Extended {{Kalman}} Filter Algorithm: A Pilot Study},
   shorttitle = {Estimation of the {{3D}} Kinematics in Kayak Using an Extended {{Kalman}} Filter Algorithm},
   author = {Fohanno, V. and Colloud, F. and Begon, M. and Lacouture, P.},
@@ -161,6 +155,12 @@
   journal = {Computer Methods in Biomechanics and Biomedical Engineering},
   language = {en},
   number = {S1}
+}
+
+@article{guennebaud2010eigen,
+  title = {Eigen V3},
+  author = {Guennebaud, Ga{\"e}l and Jacob, Beno{\^i}t and Avery, Philip and Bachrach, Abraham and Barthelemy, Sebastien and others},
+  year = {2010}
 }
 
 @article{hachajRMoCapLanguagePackage2019,
@@ -206,7 +206,7 @@
   number = {843}
 }
 
-@article{jacksonImprovementsMeasuringShoulder2012a,
+@article{jacksonImprovementsMeasuringShoulder2012,
   title = {Improvements in Measuring Shoulder Joint Kinematics},
   author = {Jackson, M. and Michaud, B. and T{\'e}treault, P. and Begon, M.},
   year = {2012},
@@ -257,7 +257,7 @@
   number = {12}
 }
 
-@book{kurfessRoboticsAutomationHandbook2018a,
+@book{kurfessRoboticsAutomationHandbook2018,
   title = {Robotics and {{Automation Handbook}}},
   author = {Kurfess, Thomas R.},
   year = {2018},
@@ -267,7 +267,6 @@
[…]
