### 2021-01-08 9c547ae
Added an example of wrapping object
