### 2021-01-08 9472493
Wrapping object are now shared_obj
