+++ src/Muscles/PathModifiers.cpp
@@ -8,7 +8,7 @@
 #include "Muscles/WrappingCylinder.h"
 
 biorbd::muscles::PathModifiers::PathModifiers() :
-    m_obj(std::make_shared<std::vector<biorbd::utils::Vector3d>>()),
+    m_obj(std::make_shared<std::vector<std::shared_ptr<biorbd::utils::Vector3d>>>()),
     m_nbWraps(std::make_shared<unsigned int>(0)),
     m_nbVia(std::make_shared<unsigned int>(0)),
     m_totalObjects(std::make_shared<unsigned int>(0))
@@ -27,7 +27,7 @@ void biorbd::muscles::PathModifiers::DeepCopy(const biorbd::muscles::PathModifie
 {
     m_obj->resize(other.m_obj->size());
     for (unsigned int i=0; i<other.m_obj->size(); ++i)
-        (*m_obj)[i] = (*other.m_obj)[i].DeepCopy();
+        *(*m_obj)[i] = (*other.m_obj)[i]->DeepCopy();
     *m_nbWraps = *other.m_nbWraps;
     *m_nbVia = *other.m_nbVia;
     *m_totalObjects = *other.m_totalObjects;
@@ -40,17 +40,20 @@ void biorbd::muscles::PathModifiers::addPathChanger(
     // Add a muscle to the pool of muscle depending on type
     if (object.typeOfNode() == biorbd::utils::NODE_TYPE::WRAPPING_SPHERE){
         biorbd::utils::Error::check(*m_nbVia == 0, "Cannot mix via points and wrapping objects yet");
-        m_obj->push_back(biorbd::muscles::WrappingSphere(static_cast<biorbd::muscles::WrappingSphere&> (object)));
+        m_obj->push_back(std::make_shared<biorbd::muscles::WrappingSphere>(
+                             static_cast<biorbd::muscles::WrappingSphere&> (object)));
         ++*m_nbWraps;
     }
     else if (object.typeOfNode() == biorbd::utils::NODE_TYPE::WRAPPING_CYLINDER){
         biorbd::utils::Error::check(*m_nbVia == 0, "Cannot mix via points and wrapping objects yet");
-        m_obj->push_back(biorbd::muscles::WrappingCylinder(dynamic_cast <biorbd::muscles::WrappingCylinder&> (object)));
+        m_obj->push_back(std::make_shared<biorbd::muscles::WrappingCylinder>(
+                             dynamic_cast <biorbd::muscles::WrappingCylinder&> (object)));
         ++*m_nbWraps;
     }
     else if (object.typeOfNode() == biorbd::utils::NODE_TYPE::VIA_POINT){
         biorbd::utils::Error::check(*m_nbWraps == 0, "Cannot mix via points and wrapping objects yet");
-        m_obj->push_back(biorbd::muscles::ViaPoint(dynamic_cast <biorbd::muscles::ViaPoint&> (object)));
+        m_obj->push_back(std::make_shared<biorbd::muscles::ViaPoint>(
+                             dynamic_cast <biorbd::muscles::ViaPoint&> (object)));
         ++*m_nbVia;
     }
     else
@@ -72,16 +75,16 @@ unsigned int biorbd::muscles::PathModifiers::nbObjects() const
     return *m_totalObjects;
 }
 
-biorbd::utils::Vector3d &biorbd::muscles::PathModifiers::object(unsigned int idx)
+biorbd::utils::Vector3d& biorbd::muscles::PathModifiers::object(unsigned int idx)
 {
     biorbd::utils::Error::check(idx<nbObjects(), "Idx asked is higher than number of wrapping objects");
-    return (*m_obj)[idx];
+    return *(*m_obj)[idx];
 }
 
 
[…]
