### 2021-01-08 05f02b3
Fixed computation of muscle length with cylinder wrapping
