+++ test/test_muscles.cpp
@@ -1549,15 +1549,17 @@ TEST(WrappingCylinder, unitTest)
         biorbd::utils::RotoTrans rt(
             biorbd::utils::Vector3d(1., 1., 1.), biorbd::utils::Vector3d(1., 1., 1.), "xyz");
 
-        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true);
+        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1., true);
         SCALAR_TO_DOUBLE(diameter, wrappingCylinder.diameter());
         EXPECT_NEAR(diameter, 0.5, requiredPrecision);
+        SCALAR_TO_DOUBLE(length, wrappingCylinder.length());
+        EXPECT_NEAR(length, 1., requiredPrecision);
     }
     {
         biorbd::utils::RotoTrans rt(
             biorbd::utils::Vector3d(1., 1., 1.), biorbd::utils::Vector3d(1., 1., 1.), "xyz");
 
-        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true, "name", "parentName");
+        biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1., true, "name", "parentName");
         EXPECT_STREQ(wrappingCylinder.parent().c_str(), "parentName");
     }
     {
@@ -1595,7 +1597,7 @@ TEST(WrappingCylinder, deepCopy)
     biorbd::utils::RotoTrans rt(
         biorbd::utils::Vector3d(1, 1, 1), biorbd::utils::Vector3d(1, 1, 1), "xyz");
 
-    biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, true);
+    biorbd::muscles::WrappingCylinder wrappingCylinder(rt, 0.5, 1., true);
 
     biorbd::muscles::WrappingCylinder shallowCopy(wrappingCylinder);
     biorbd::muscles::WrappingCylinder deepCopyNow(wrappingCylinder.DeepCopy());
@@ -1611,10 +1613,20 @@ TEST(WrappingCylinder, deepCopy)
         EXPECT_NEAR(shallowCopyDiameter, 0.5, requiredPrecision);
         EXPECT_NEAR(deepCopyNowDiameter, 0.5, requiredPrecision);
         EXPECT_NEAR(deepCopyLaterDiameter, 0.5, requiredPrecision);
+
+        SCALAR_TO_DOUBLE(length, wrappingCylinder.length());
+        SCALAR_TO_DOUBLE(shallowCopyLength, shallowCopy.length());
+        SCALAR_TO_DOUBLE(deepCopyNowLength, deepCopyNow.length());
+        SCALAR_TO_DOUBLE(deepCopyLaterlength, deepCopyLater.length());
+        EXPECT_NEAR(length, 1.0, requiredPrecision);
+        EXPECT_NEAR(shallowCopyLength, 1.0, requiredPrecision);
+        EXPECT_NEAR(deepCopyNowLength, 1.0, requiredPrecision);
+        EXPECT_NEAR(deepCopyLaterlength, 1.0, requiredPrecision);
     }
 
     // Set new diameter
     wrappingCylinder.setDiameter(2.0);
+    wrappingCylinder.setLength(2.5);
 
     // Check values
     {
@@ -1624,8 +1636,17 @@ TEST(WrappingCylinder, deepCopy)
         SCALAR_TO_DOUBLE(deepCopyLaterDiameter, deepCopyLater.diameter());
         EXPECT_NEAR(diameter, 2., requiredPrecision);
[…]
