+++ binding/biorbd_muscles.i.in
@@ -17,6 +17,8 @@
 #include "Muscles/State.h"
 #include "Muscles/StateDynamics.h"
 #include "Muscles/StateDynamicsBuchanan.h"
+#include "Muscles/WrappingObject.h"
+#include "Muscles/WrappingCylinder.h"
 #ifdef MODULE_STATIC_OPTIM
 #include "Muscles/StaticOptimization.h"
 #endif
@@ -106,5 +108,7 @@ namespace std {
 %include "@CMAKE_SOURCE_DIR@/include/Muscles/State.h"
 %include "@CMAKE_SOURCE_DIR@/include/Muscles/StateDynamics.h"
 %include "@CMAKE_SOURCE_DIR@/include/Muscles/StateDynamicsBuchanan.h"
+%include "@CMAKE_SOURCE_DIR@/include/Muscles/WrappingObject.h"
+%include "@CMAKE_SOURCE_DIR@/include/Muscles/WrappingCylinder.h"
 @SWIG_STATIC_OPTIMIZATION_INCLUDE_COMMAND@
 
