### 2021-01-16 #2562 @arfon
@pariterre - could you take a look at the `@eigenweb` citation in your paper - I'm not sure it resolves to anything in your BibTeX file?
