### 2021-01-16 #2562 @arfon
Also, can I just check the authorship list here? In your paper you say "A huge thanks to Ariane Dang for her patience and contribution to writing the tests for biorbd!". This author looks to be the `#2` [contributor to this package](https://github.com/pyomeca/biorbd/graphs/contributors) but doesn't look to be an author here.

Could you say a little more about this please?
