### 2021-01-18 3b67441
Fixed eigen reference in the paper
