### 2021-01-18 #2562 @pariterre
Dear @arfon

> @pariterre - could you take a look at the @eigenweb citation in your paper - I'm not sure it resolves to anything in your BibTeX file?

That is right, I don't know why though.. It is now fixed!

> Also, can I just check the authorship list here? In your paper you say "A huge thanks to Ariane Dang for her patience and contribution to writing the tests for biorbd!". This author looks to be the #2 contributor to this package but doesn't look to be an author here.
>
> Could you say a little more about this please?

Ariane was a colleague of mine at the lab where I study (she now works elsewhere). She is an amazing person and I really like her very much. That said, she was mainly applying recipes that I was preparing for her (for instance, I was preparing a test, and she was copying it for all the relevant classes). I honestly have no problem adding her to the authorship, but I fell it would be kind of overkill.. As for her, when we discussed her position in the paper, she seemed totally fine with being only thanked. We can ping her if you want to confirm this information
