### 2021-01-19 #2562 @pariterre
@Dangzilla
I hope you are doing fine! We miss you at the lab :D

The reviewers for the paper on biorbd would like to confirm that you feel good about the fact that you are in the `acknowledgement` section instead of in the authors list. As you will see in earlier post, I told them that we basically worked together, and that most of the tests were written from a prepared recipe.

Please fell free to answer entirely as you fell. So, if you fell that you should be part of the author list, please say so :)
