### 2021-01-19 #2562 @pariterre
Hi @arfon
I uploaded the 1.4.1 version (created after your last comment) of biorbd on Zenodo. The DOI is 10.5281/zenodo.4450303. I am not sure how I update this thread with the DOI though..
I made sure that the names and the paper title fit :)
Thanks!
