### 2021-01-19 #2562 @arfon
@pariterre - I made a small number of formatting fixes in https://github.com/pyomeca/biorbd/pull/193
