### 2021-01-19 #2562 @arfon
> It is now merged! Did I go too fast with the version on Zenodo?

No, it's fine. We don't need these last fixes to the paper in the Zenodo archive.
