### 2021-01-19 #2562 @arfon
@trallard, @abhishektha - many thanks for your reviews here and to @trallard for editing this submission too! JOSS relies upon the volunteer efforts of folks like yourselves and we simply wouldn't be able to do this without you ✨

@pariterre your paper is now accepted and published in JOSS :zap::rocket::boom:
